# Conflicts:
#	cpp/include/egttools/Distributions.h
#	cpp/src/egttools.cpp
#	cpp/src/egttools/Distributions.cpp
#	cpp/src/version.h
#	pybind11
#	src/egttools/analytical/sed_analytical.py

+++ cpp/include/egttools/finite_populations/games/AbstractNPlayerGame.hpp
@@ -85,7 +85,7 @@ namespace egttools::FinitePopulations {
 
         [[nodiscard]] virtual size_t group_size() const;
 
-        [[nodiscard]] virtual size_t nb_group_configurations() const;
+        [[nodiscard]] virtual int64_t nb_group_configurations() const;
 
         /**
          * @return Returns a small description of the game.
